Make base types more consistent
Changes to base types:
* make numeric a union
* add new atomic and vector unions
* remove factor and data.frame S3 classes
* add expression base type

And includes a few quality of life improvements:

* Method dispatch failure now reports correct classes
* Methods created from union now have own names
* More explicit error when parent is a union

Fixes #77. Fixes #94. Built on top of #134.

+++ tests/testthat/helper.R
@@ -1,5 +1,5 @@
 text <- new_class("text", parent = "character", constructor = function(text = character()) new_object(.data = text))
-number <- new_class("number", parent = "numeric", constructor = function(x) new_object(.data = x))
+number <- new_class("number", parent = "double", constructor = function(x) new_object(.data = x))
 
 range <- new_class("range",
   constructor = function(start, end) {
@@ -11,11 +11,11 @@ range <- new_class("range",
     }
   },
   properties = list(
-    start = "numeric",
-    end = "numeric",
+    start = "double",
+    end = "double",
     new_property(
       name = "length",
-      class = "numeric",
+      class = "double",
       getter = function(x) x@end - x@start,
       setter = function(x, value) {
         x@end <- x@start + value
